requirements.txt
apache-airflow==2.0.0
pandas
requests
sqlalchemy
psycopg2-binary
